// packages/fb-components/rich_text/types.ts
export enum InlineEntryHyperlinkType {
  MentionUser = 'mention_user',
  MentionRole = 'mention_role',
  Channel = 'channel',
  Topic = 'topic',
}

export enum EmbeddedAssetType {
  Image = 'image',
  Video = 'video',
}

export enum EmbeddedResourceType {
  Question = 'question',
}

export enum TextMentionSign {
  User = '!',
  Role = '&',
}

// 导出 default
export default {}